+++ .gitignore
@@ -2,3 +2,4 @@
 data.html
 data.txt
 output
+relevant_data
