+++ requirements.txt
@@ -1,4 +1,54 @@
-pandas==1.3.5
-spacy==3.7.1
-ipython==8.16.1
-spacy-transformers==1.3.1
+annotated-types==0.6.0
+blis==0.7.11
+catalogue==2.0.10
+certifi==2023.7.22
+charset-normalizer==3.3.2
+click==8.1.7
+cloudpathlib==0.16.0
+colorama==0.4.6
+confection==0.1.3
+cupy-cuda116==10.6.0
+cymem==2.0.8
+en-core-web-sm==3.7.0
+fastrlock==0.8.2
+filelock==3.13.1
+fsspec==2023.10.0
+huggingface-hub==0.17.3
+idna==3.4
+Jinja2==3.1.2
+langcodes==3.3.0
+MarkupSafe==2.1.3
+mpmath==1.3.0
+murmurhash==1.0.10
+networkx==3.1
+numpy==1.23.5
+packaging==23.2
+Pillow==10.1.0
+preshed==3.0.9
+pydantic==2.5.0
+pydantic_core==2.14.1
+PyYAML==6.0.1
+regex==2023.10.3
+requests==2.31.0
+safetensors==0.4.0
+smart-open==6.4.0
+spacy==3.7.2
+spacy-alignments==0.9.1
+spacy-legacy==3.0.12
+spacy-loggers==1.0.5
+spacy-lookups-data==1.0.5
+spacy-transformers==1.3.3
+srsly==2.4.8
+sympy==1.12
+thinc==8.2.1
+tokenizers==0.14.1
+torch==1.12.1+cu116
+torchaudio==0.12.1+cu116
+torchvision==0.13.1+cu116
+tqdm==4.66.1
+transformers==4.35.0
+typer==0.9.0
+typing_extensions==4.8.0
+urllib3==2.1.0
+wasabi==1.1.2
+weasel==0.3.4
